optimize dependency installation of user code

+++ entrypoint.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
-pip3 install -r /user_code/requirements.txt
 
-exec "$@"
+if [ ! -f dependencies.txt ] || ! cmp -s dependencies.txt /user_code/requirements.txt; then
+    pip3 install -r /user_code/requirements.txt
+    cp /user_code/requirements.txt dependencies.txt
+else
+    exec "$@"
+fi
